LPS-94301 dynamic-data-mapping - classes in module that are not exported should be in internal package
