+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-web/src/main/java/com/liferay/dynamic/data/mapping/web/internal/configuration/DDMWebConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.web.configuration;
+package com.liferay.dynamic.data.mapping.web.internal.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
 
@@ -23,7 +23,7 @@ import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClass
  */
 @ExtendedObjectClassDefinition(category = "dynamic-data-mapping")
 @Meta.OCD(
-	id = "com.liferay.dynamic.data.mapping.web.configuration.DDMWebConfiguration",
+	id = "com.liferay.dynamic.data.mapping.web.internal.configuration.DDMWebConfiguration",
 	localization = "content/Language", name = "ddmweb-configuration-name"
 )
 public interface DDMWebConfiguration {
